+++ src/main/java/com/example/solidconnection/admin/dto/LanguageTestScoreUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.admin.dto;
 
-import com.example.solidconnection.common.validation.annotation.RejectedReasonRequired;
+import com.example.solidconnection.application.dto.validation.RejectedReasonRequired;
 import com.example.solidconnection.university.domain.LanguageTestType;
 import com.example.solidconnection.application.domain.VerifyStatus;
 import jakarta.validation.constraints.NotNull;
